+++ structr-ui/src/test/java/org/structr/test/web/StructrUiTest.java
@@ -77,9 +77,13 @@ public abstract class StructrUiTest {
 	protected final String htmlUrl     = "/structr/html";
 	protected final String wsUrl       = "/structr/ws";
 	protected String baseUri           = null;
+	private boolean first              = true;
 
 	@BeforeClass(alwaysRun = true)
 	public void setup() {
+
+		// for parallel test execution, add random startup delay
+		try { Thread.sleep(new Random().nextInt(5000)); } catch (Throwable t) {}
 
 		final long timestamp = System.nanoTime();
 
@@ -140,25 +144,30 @@ public abstract class StructrUiTest {
 		System.out.println("######################################################################################");
 	}
 
-	@AfterMethod
+	@BeforeMethod()
 	public void cleanDatabase() {
 
-		try (final Tx tx = app.tx()) {
+		if (!first) {
 
-			// delete everything
-			Services.getInstance().getService(NodeService.class).getDatabaseService().cleanDatabase();
+			try (final Tx tx = app.tx()) {
 
-			FlushCachesCommand.flushAll();
+				// delete everything
+				Services.getInstance().getService(NodeService.class).getDatabaseService().cleanDatabase();
 
-			SchemaService.ensureBuiltinTypesExist(app);
+				FlushCachesCommand.flushAll();
 
-			tx.success();
+				SchemaService.ensureBuiltinTypesExist(app);
 
-		} catch (Throwable t) {
+				tx.success();
 
-			t.printStackTrace();
-			logger.error("Exception while trying to clean database: {}", t.getMessage());
+			} catch (Throwable t) {
+
+				t.printStackTrace();
+				logger.error("Exception while trying to clean database: {}", t.getMessage());
+			}
 		}
+
+		first = false;
 	}
 
 	@AfterClass(alwaysRun = true)
